+++ R/bdotsBoot.R
@@ -108,37 +108,45 @@
 #'
 #' @import data.table
 #' @export
-
 bdotsBoot <- function(formula,
                       bdObj,
                       Niter = 1000,
                       alpha = 0.05,
                       padj = "oleson",
                       cores = 0, ...) {
-
+  
   if (cores < 1) cores <- detectCores()/2
-
+  
   if (any(bdObj[['fitCode']] == 6)) {
     warning("Some observations had NULL gnls fits. These and their pairs will be removed")
     bdObj <- bdRemove(bdObj, fitCode = 6, removePairs = TRUE)
   }
-
+  
   prs <- bootParser(formula, bdObj)
   bdObj <- bootGroupSubset(prs, bdObj)
   innerDiff <- prs[["innerDiff"]] # unnamed char vec
   outerDiff <- prs[["outerDiff"]] # unnamed char vec
-
+  
   curveGrps <- setNames(prs[['subargs']], prs[['subnames']])
-
   curveFun <- makeCurveFun(bdObj)
-
-  curveList <- curveBooter(bdObj,
-                           outerDiff = outerDiff,
-                           innerDiff = innerDiff,
-                           N.iter = Niter,
-                           curveFun = curveFun)
+  
+  ## Next, we want to get a bootstrapped distribution for each of the groups
+  splitGroups <- split(bdObj, by = c(innerDiff, outerDiff)) # ok even if null
+  
+  ## Make this parallel (maybe later)
+  if (Sys.info()['sysname'] == "Darwin") {
+    cl <- makePSOCKcluster(cores, setup_strategy = "sequential")
+  } else {
+    cl <- makePSOCKcluster(cores)
+  }
+  invisible(clusterEvalQ(cl, {library(bdots)}))
+  groupDists <- parLapply(cl, splitGroups, getBootDist, b = Niter)
+  
+  ## Next, we need to make inner/outer diff list
+  # (ideally matching old bdots, at least for now)
+  curveList <- createCurveList(groupDists, prs, splitGroups)
   ip <- curveList[['diff']][['paired']]
-
+  
   res <- alphaAdjust(curveList, padj, alpha, cores)
   pval <- res[['pval']]
   rho <- res[['rho']]
@@ -147,7 +155,7 @@ bdotsBoot <- function(formula,
   time <- attr(bdObj, "time")
   sigTime <- bucket(adjpval <= alpha, time)
   dod <- ifelse(is.null(innerDiff), FALSE, TRUE)
-
+  
   ## maybe consider keeping ancillary data in list, i.e., res.
   structure(class = "bdotsBootObj",
             .Data = list(curveList = curveList,
@@ -165,11 +173,4 @@ bdotsBoot <- function(formula,
             call = match.call(),
             bdObjAttr = attributes(bdObj))
 }
-
-
-
-
-
-
-